Separate log based test from PeerSpec to have a cleaner output
