+++ eclair-core/src/test/scala/fr/acinq/eclair/TestkitBaseClass.scala
@@ -49,25 +49,3 @@ abstract class TestkitBaseClass extends TestKit(ActorSystem("test")) with fixtur
   }
 
 }
-
-abstract class LoggingTestkitBaseClass extends TestKit(ActorSystem("test", ConfigFactory.parseString("""akka.loggers = ["akka.testkit.TestEventListener"]"""))) with fixture.FunSuiteLike with BeforeAndAfterEach with BeforeAndAfterAll {
-
-  override def beforeAll {
-    Globals.blockCount.set(400000)
-    Globals.feeratesPerKw.set(FeeratesPerKw.single(TestConstants.feeratePerKw))
-  }
-
-  override def afterEach() {
-    system.actorSelection(system / "*") ! PoisonPill
-    intercept[ActorNotFound] {
-      import scala.concurrent.duration._
-      Await.result(system.actorSelection(system / "*").resolveOne(42 days), 42 days)
-    }
-  }
-
-  override def afterAll {
-    TestKit.shutdownActorSystem(system)
-    Globals.feeratesPerKw.set(FeeratesPerKw.single(1))
-  }
-
-}
